third_party/blink/web_tests/http/tests/devtools/elements/shadow/elements-panel-shadow-selection-on-refresh-1-expected.txt

Tests that elements panel preserves selected shadow DOM node on page refresh.


Running: setup

Running: testOpenShadowRoot
Page reloaded.
Selected node: '#shadow-root (open)'

Running: testClosedShadowRoot
Page reloaded.
Selected node: '#shadow-root (closed)'
